requirements.txt
beautifulsoup4==4.13.5
deepgram_sdk==2.12.0
elevenlabs==0.2.27
Flask==3.1.2
protobuf==6.32.0
pyaudio==0.2.14
pygame==2.5.2
Requests==2.32.5
rhasspy_silence==0.4.0

blinker==1.9.0
click==8.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3

python-dotenv==1.0.0
marshmallow==3.19.0
google-genai==0.2.0